requirements.txt
google-genai>=0.14.0
psycopg2-binary>=2.9
typer[all]
rich
ruff